dev: consolidate diff and commit handling for model commit and load
